Get rid of remaining casts to "unsigned long" by using "%zu"
Still no changes, just avoid ugly casts by using the correct formatting
string instead.

+++ src/classes/MModule.cpp
@@ -733,8 +733,7 @@ MModule::ListAvailableModules(const String& interfaceName)
       }
    }
 
-   wxLogTrace(M_TRACE_MODULES, _T("\t%lu modules found."),
-              (unsigned long)count);
+   wxLogTrace(M_TRACE_MODULES, _T("\t%zu modules found."), count);
 
    listing->SetCount(count);
